+++ CHANGELOG.md
@@ -1,7 +1,19 @@
+# 0.3.0
+- Add Dreamer wrapper for recurrent state space modeling, `models.RSSM`
+- Add classic gymnasium environments
+- Add general purpose models `models.MLP`, `models.MLPEncoder`, `models.MLPDecoder`, and `models.CNNDecoder`
+- Add initial environment loop for testing
+- Add recurrent model `models.RecurrentModel` with layer `models.LayerNormGRU` and corresponding unit test
+- Add uniform mix `distributions.uniform_mix` and symlog `distributions.symlog`
+- Changes to `models.CNNEncoder` final layer
+- Display `__init__` methods in documentation
+- Group ordering of classes and functions in documentation
+
 # 0.2.0
-- Add convolutional encoder `modules.CNNEncoder` and channel layer normalization `modules.ChannelNorm`
+- Add convolutional encoder `models.CNNEncoder` and channel layer normalization `models.ChannelNorm`
 - Add tests for `models` module
 - Further refine linting and add module docstrings
+- `README.md` formatting fix
 
 # 0.1.0
 - Boilerplate, documentation with `Sphinx`, linting with `Ruff`, unit testing